+++ mpich2/test/mpi/threads/pt2pt/multisend3.c
@@ -30,7 +30,7 @@ static int nthreads = -1;
 MTEST_THREAD_RETURN_TYPE run_test_send(void *arg)
 {
     int    cnt, j, *buf, wsize;
-    int    thread_num = (int)arg;
+    int    thread_num = (int)(size_t)arg;
     double t;
     static MPI_Request r[MAX_NTHREAD];
 
@@ -119,7 +119,7 @@ int main(int argc, char ** argv)
     if (rank == 0) {
 	nthreads = nprocs - 1;
 	for (i=1; i<nprocs; i++) 
-	    MTest_Start_thread( run_test_send,  (void *)i );
+	    MTest_Start_thread( run_test_send,  (void *)(size_t)i );
 
 	MTest_Join_threads( );
     }
